+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fixed a typo in D_regularized manual
 * Added difference between two dependent correlations -function (diff_two_dep_cors) which enables simultaneous estimation and testing of Cohen's q under variable dependency
 * Added possibility to use manually constructed regularization and estimation datasets, supplied as a list of two dataframes to "data"-argument in D_regularized
+* Improved the output of diff_two_dep_cors and included an agument for missing data ("ML")
 
 # multid 0.7.1